C++ backend for Treap

+++ pydatastructs/utils/_backend/cpp/TreeNode.hpp
@@ -19,7 +19,7 @@ typedef struct {
     long size;
     long color;
     bool isCartesianTreeNode;
-    long priority;
+    double priority;
 } TreeNode;
 
 static void TreeNode_dealloc(TreeNode *self) {
